test: add integration tests and finalize business logic adjustments
