+++ src/Ambev.DeveloperEvaluation.WebApi/Features/Sales/CancelItemSale/CancelItemSaleProfile.cs
@@ -7,7 +7,9 @@ public class CancelItemSaleProfile : Profile
 {
   public CancelItemSaleProfile()
   {
-    CreateMap<CancelItemSaleRequest, CancelItemCommand>();
+    CreateMap<CancelItemSaleRequest, CancelItemCommand>()
+        .ForMember(dest => dest.SaleId, opt => opt.MapFrom(src => src.SaleId))
+        .ForMember(dest => dest.ItemId, opt => opt.MapFrom(src => src.Id));
     CreateMap<CancelItemResult, CancelItemSaleResponse>();
   }
 }
