+++ src/Ambev.DeveloperEvaluation.WebApi/Features/Sales/CancelItemSale/CancelItemSaleRequest.cs
@@ -2,5 +2,6 @@ namespace Ambev.DeveloperEvaluation.WebApi.Features.Sales.CancelItemSale;
 
 public class CancelItemSaleRequest
 {
+  public Guid SaleId { get; set; }
   public Guid Id { get; set; }
 }
