+++ src/Ambev.DeveloperEvaluation.WebApi/Features/Sales/CancelSale/CancelSaleProfile.cs
@@ -5,9 +5,10 @@ namespace Ambev.DeveloperEvaluation.WebApi.Features.Sales.CancelSale;
 
 public class CancelSaleProfile : Profile
 {
-  public CancelSaleProfile()
-  {
-    CreateMap<CancelSaleRequest, CancelSaleCommand>();
-    CreateMap<CancelSaleResult, CancelSaleResponse>();
-  }
+    public CancelSaleProfile()
+    {
+        CreateMap<CancelSaleRequest, CancelSaleCommand>();
+        CreateMap<CancelSaleResult, CancelSaleResponse>();
+
+    }
 }
