+++ src/Ambev.DeveloperEvaluation.WebApi/Features/Sales/CancelSale/CancelSaleRequest.cs
@@ -2,5 +2,6 @@ namespace Ambev.DeveloperEvaluation.WebApi.Features.Sales.CancelSale;
 
 public class CancelSaleRequest
 {
-  public Guid Id { get; set; }
+    public Guid Id { get; set; }
+    public string CancellationReason { get; set; }
 }
