+++ src/Ambev.DeveloperEvaluation.WebApi/Features/Sales/GetSale/GetSaleProfile.cs
@@ -1,14 +1,15 @@
 using Ambev.DeveloperEvaluation.Application.Sales.GetSale;
-using Ambev.DeveloperEvaluation.WebApi.Features.Sales.GetSale;
 using AutoMapper;
 
-namespace Ambev.DeveloperEvaluation.WebApi.Features.Sale.GetSale;
+namespace Ambev.DeveloperEvaluation.WebApi.Features.Sales.GetSale;
 
 public class GetSaleProfile : Profile
 {
     public GetSaleProfile()
     {
         CreateMap<GetSaleRequest, GetSaleCommand>();
+        CreateMap<Guid, GetSaleCommand>()
+            .ConstructUsing(id => new GetSaleCommand() { Id = id });
         CreateMap<GetSaleResult, GetSaleResponse>();
         CreateMap<GetSaleItemResult, GetSaleItemResponse>();
     }
